+++ models/ingressos.py
@@ -13,7 +13,7 @@ class Ingressos(db.Model):
     codigo = db.Column(db.Integer, primary_key=True, nullable=True, unique=True)
     tipo = db.Column(db.String(20))
     status = db.Column(db.String(20))
-    codusuario = db.Column(db.Integer(5))
+    codusuario = db.Column(db.Integer)
 
     def __init__(self, codigo,tipo,status,codusuario):
         self.codigo= codigo
